Split certificate.rs and lib.rs at clean boundaries to clear max-lines
