+++ crates/almide-mir/src/lib.rs
@@ -497,985 +497,5 @@ pub enum Op {
     SetLocal { local: ValueId, src: ValueId },
 }
 
-/// The closed set of primitive-floor operations (the trusted, wasm-spec-faithful
-/// surface the self-hosted runtime is written over).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum PrimKind {
-    /// Reinterpret a heap handle (i32 pointer) as an i64 address value — the
-    /// String/List→Int bridge so all address math is `Int` `IntBinOp`.
-    Handle,
-    /// Load `width` bytes (1/4/8) at a computed i64 address, zero-extended to i64.
-    Load { width: u8 },
-    /// Load a 4-byte i32 HANDLE at a computed i64 address — UNLIKE `Load { width: 4 }`, the
-    /// result keeps the `Ptr` (i32) repr (no i64 zero-extend), so it IS a heap handle a caller
-    /// can pass to a String/List consumer. The bridge for extracting a heap element from a slot
-    /// (a `match Some(s)` payload / a `List[String]` element). A borrowed alias — no ownership.
-    LoadHandle,
-    /// Store the low `width` bytes (1/4/8) of an i64 value at a computed i64 address.
-    Store { width: u8 },
-    /// Bounds-checked element ADDRESS for a direct `xs[i]` index — `args = [list_handle,
-    /// index]` (both i64-uniform: the handle reinterpreted to an address, the index a scalar
-    /// i64), dst = the i64 element-slot address `list + LIST_HEADER + idx*ELEM_SIZE`. Renders
-    /// `(call $elem_addr ...)` (the SAME preamble helper v0's `$list_get`/`$list_set` use), so a
-    /// negative or `>= cap` index TRAPS (the controlled-halt bounds wall) instead of reading
-    /// outside the block — v0's `a[i]` likewise halts on OOB (it prints `index out of bounds`
-    /// and exits 1; this traps). For an in-bounds index the loaded element byte-matches v0. A
-    /// scalar address computation, no ownership (a no-op in verify_ownership like every Prim).
-    ElemAddr,
-    /// Abort: write the String-block message to STDERR and proc_exit(1) — the
-    /// self-host arm of the §13 termination convention (math.pow negative
-    /// exponent, int.rotate nonpositive width). Never returns.
-    Die,
-    /// `process.exit(code)` — the WASI `proc_exit` host call with a USER exit
-    /// code (`args = [code]`, i64 wrapped to i32; no message line, unlike
-    /// [`PrimKind::Die`]'s fixed exit-1 + stderr). Never returns; carries no
-    /// ownership event and no capability of its own (the frontend's E006
-    /// already forces the calling fn to be `effect`). #782: the assert/T18
-    /// desugar's `process.exit(1)` statement rode the retired v0 emitter —
-    /// this is its v1 floor.
-    ProcExit,
-    /// The `fd_write` WASI host call — `args = [fd, iov, count, nwritten]`, dst = the
-    /// i64 errno. A sandbox exit; carries [`Capability::Stdout`].
-    FdWrite,
-    /// The `random_get` WASI host call — `args = [buf, buf_len]`, dst = the i64 errno;
-    /// fills `buf_len` bytes at `buf` with host entropy. The second sandbox exit; reached
-    /// only by the self-hosted `random.int`. Carries [`Capability::Entropy`] (the
-    /// cap_witness counts it exactly like `FdWrite` → Stdout), so a function using it is
-    /// caps-verified ONLY if it declares Entropy — never accept-but-unsafe.
-    RandomGet,
-    /// The `clock_time_get` WASI host call — `args = [clock_id, precision, time_ptr]`, dst =
-    /// the i64 errno; writes the current clock value (nanoseconds) as an i64 at `time_ptr`.
-    /// A SCALAR-dst sandbox exit (like [`RandomGet`] — NO heap result, NO ownership event),
-    /// reached only by the self-hosted `env.unix_timestamp` (which reads `time_ptr` and
-    /// divides by 1e9 to seconds). Carries [`Capability::Clock`] — a DISTINCT capability
-    /// (a clock read is neither a filesystem nor an entropy effect; the cap_witness counts
-    /// it exactly like `RandomGet` → Entropy), so a function using it is caps-verified ONLY
-    /// if it declares Clock — never accept-but-unsafe. NON-DETERMINISTIC (no byte-match).
-    ClockTimeGet,
-    /// The `args_sizes_get` + `args_get` WASI host calls, packaged as ONE high-level
-    /// HEAP-RESULT prim — no args, dst = a fresh OWNED `List[String]` of the program
-    /// arguments `argv[1..]` (SKIP argv[0], matching native `env.args`). Each element
-    /// is a canonical Almide String copied from the NUL-terminated argv C-string. The
-    /// third sandbox exit, reached only by the self-hosted `env.args`. Carries
-    /// [`Capability::CliArgs`] (the cap_witness counts it exactly like `RandomGet` →
-    /// Entropy), so a function using it is caps-verified ONLY if it declares CliArgs —
-    /// never accept-but-unsafe. Its dst is a heap Ptr (like `LoadHandle`), so the
-    /// ownership certificate emits an `i` (alloc) for it, balanced by the caller's
-    /// scope-end drop (a recursive `DropListStr` over the owned element Strings).
-    ArgsGetList,
-    /// The SAME WASI args floor as [`ArgsGetList`] but INCLUDING argv[0] (the program
-    /// path) — `process.args()` = native `std::env::args()`. Renders as
-    /// `(call $args_get_list (i32.const 0))` (the one parameterized bridge, skip=0);
-    /// same fresh OWNED `List[String]` dst, same [`Capability::CliArgs`] accounting.
-    ArgsGetListFull,
-    /// The WASI `environ_sizes_get` + `environ_get` lookup, packaged as ONE high-level
-    /// HEAP-RESULT prim — `args = [name]` (a BORROWED `String` handle), dst = a fresh
-    /// OWNED `Option[String]`: a 0-slot block (none) or a 1-slot block whose @12 holds
-    /// the owned value String (some) — the `materialize_opt_str_some` layout, so the
-    /// caller's `match`/`??`/`DropListStr` machinery handles it identically to a
-    /// self-host-built Option. Scans the `KEY=VALUE\0` environ entries for `name`
-    /// followed by `=` (byte-exact, first hit wins) — native `std::env::var(name).ok()`
-    /// is the oracle (C-133; the runner passes the host env through
-    /// `wasmtime -S inherit-env=y`). Reached only by the self-hosted `env.get`.
-    /// Carries [`Capability::CliArgs`] — the Env effect-profile's canonical capability
-    /// (reading the process's initial environment, the same class as argv; the profile
-    /// map `"Env" => CliArgs` already binds them). Its dst is a heap Ptr (like
-    /// [`ArgsGetList`]), so the ownership certificate emits an `i` (alloc) for it,
-    /// balanced by the caller's scope-end drop (the flat `DropListStr` frees the owned
-    /// payload String, if any, then the block) or a heap-return move-out.
-    EnvGet,
-    /// The WASI `fd_read`-from-stdin line-read sequence, packaged as ONE high-level HEAP-RESULT
-    /// prim — no args, dst = a fresh OWNED canonical `String` of ONE line of standard input.
-    /// Reads fd 0 BYTE-BY-BYTE (so it never over-reads past the newline — a later
-    /// `read_n_bytes` of the body still sees the right stream) until a `\n` (excluded from the
-    /// result) or EOF, then strips a trailing `\r` (matching native
-    /// `read_line().trim_end_matches('\n').trim_end_matches('\r')`). EOF with no bytes yields the
-    /// empty String. Reached only by the self-hosted `io.read_line`. Carries [`Capability::Stdin`]
-    /// — a DISTINCT capability (reading standard input is neither a write, a filesystem, an
-    /// entropy, nor a clock effect; the cap_witness counts it exactly like `RandomGet` → Entropy),
-    /// so a function using it is caps-verified ONLY if it declares Stdin — never accept-but-unsafe.
-    /// NON-DETERMINISTIC (reads live stdin): no byte-match across runs unless stdin is fixed. Its
-    /// dst is a heap Ptr (like [`ArgsGetList`]), so the ownership certificate emits an `i` (alloc)
-    /// for it, balanced by the caller's scope-end flat `Drop` (a String owns no nested handles) or
-    /// a heap-return move-out.
-    ReadLine,
-    /// `read_n_bytes(n)` — the WASI stdin-N-bytes floor (io.read_n_bytes), the SIBLING of
-    /// [`PrimKind::ReadLine`]: `args = [n]` (an `Int`, the byte count), dst = a fresh OWNED `Bytes`
-    /// block (the same byte-buffer block layout a `String` uses, built by the preamble `$read_n_bytes`
-    /// via `$rtf_str`). Reads UP TO `n` bytes from fd 0 (stopping early at EOF). Carries
-    /// Capability::Stdin (same DISTINCT cap as ReadLine). NON-DETERMINISTIC (live stdin): no byte-match.
-    /// Its dst is a heap Ptr, so the ownership certificate emits an `i` (alloc) balanced by the caller's
-    /// scope-end flat `Drop` (a Bytes owns no nested handles) or a heap-return move-out.
-    ReadNBytes,
-    /// The WASI `path_open` + `fd_read` file-read sequence, packaged as ONE high-level
-    /// HEAP-RESULT prim — `args = [path]` (a BORROWED `String` handle), dst = a fresh
-    /// OWNED `Result[String, String]`. Opens the file at `path` (relative to the first
-    /// preopened dir, leading `/` stripped — the same absolute-path fallback the native
-    /// emit's `__resolve_path` uses) and reads its bytes: on success builds `Ok(content)`
-    /// where `content` is a canonical Almide String of the file bytes; on a path_open
-    /// error builds `Err(<message>)`. The result block is the EXACT `materialize_result_str`
-    /// layout — a 1-slot DynListStr `[rc][len@4=1][cap@8][@12 String handle][@16 tag]`
-    /// (tag 0 = Ok, 1 = Err) — so the caller's `!`/`match`/`DropListStr` machinery handles
-    /// it identically to a self-host-built `Result[String, String]`. The FOURTH sandbox
-    /// exit, reached only by the self-hosted `fs.read_text`. Carries [`Capability::FsRead`]
-    /// (the cap_witness counts it exactly like `ArgsGetList` → CliArgs), so a function using
-    /// it is caps-verified ONLY if it declares FsRead — never accept-but-unsafe. Its dst is
-    /// a heap Ptr (like `ArgsGetList`), so the ownership certificate emits an `i` (alloc) for
-    /// it, balanced by the caller's scope-end drop (the flat `DropListStr` over the one owned
-    /// payload String).
-    ReadTextFile,
-    /// The WASI `path_open(O_DIRECTORY)` + `fd_readdir` directory-listing sequence, packaged
-    /// as ONE high-level HEAP-RESULT prim — `args = [path]` (a BORROWED `String` handle), dst
-    /// = a fresh OWNED `Result[List[String], String]`. Opens the directory at `path` (same
-    /// preopen-relative resolution as [`ReadTextFile`]) and reads its entries via an
-    /// `fd_readdir` re-read-on-truncation loop, parsing each variable-length dirent record
-    /// (`d_next u64 / d_ino u64 / d_namlen u32 / d_type u8 / name[d_namlen]`), SKIPPING `.`
-    /// and `..` (WASI yields them, native `std::fs::read_dir` does not), then SORTING the names
-    /// lexicographically (to byte-match the Rust runtime's `names.sort()`), and builds
-    /// `Ok([name, …])` where the payload is a fresh owned `List[String]`. On a path_open error
-    /// it builds `Err(<message>)`. The result block is the cap-as-tag layout `[rc][len@4=1]
-    /// [cap@8=1][@12 List[String] handle][@16 tag]` (tag 0 = Ok, 1 = Err) — the SAME shape as
-    /// [`ReadTextFile`], only the @12 payload is a nested `List[String]` (so the scope-end drop
-    /// is the RECURSIVE [`StmtKind::DropResultListStr`], not the flat `DropListStr` that would
-    /// leak the inner element Strings). The FIFTH sandbox exit, reached only by the self-hosted
-    /// `fs.list_dir`. Carries [`Capability::FsRead`] (the cap_witness counts it exactly like
-    /// [`ReadTextFile`] → FsRead), so a function using it is caps-verified ONLY if it declares
-    /// FsRead — never accept-but-unsafe. Its dst is a heap Ptr (like [`ReadTextFile`]), so the
-    /// ownership certificate emits an `i` (alloc) for it, balanced by the caller's scope-end
-    /// recursive drop (or a heap-return move-out).
-    ReadDir,
-    /// The WASI `path_open(O_CREAT|O_TRUNC)` + `fd_write` file-WRITE sequence, packaged as ONE
-    /// high-level HEAP-RESULT prim — `args = [path, content]` (both BORROWED `String` handles,
-    /// the caller still owns them), dst = a fresh OWNED `Result[Unit, String]`. Opens (creating +
-    /// truncating) the file at `path` (relative to the first preopened dir, leading `/` stripped —
-    /// the same resolution [`ReadTextFile`] uses) and writes `content`'s bytes via `fd_write`: on
-    /// success builds `Ok(())`, on a path_open / fd_write error builds `Err(<message>)`. The result
-    /// block reuses the cap-as-tag layout `[rc][len@4][cap@8][@12][@16 tag]` (tag 0 = Ok, 1 = Err),
-    /// but DIVERGES from [`ReadTextFile`] in the Ok arm: a `Unit` payload owns NO String, so Ok is
-    /// built with `len@4 = 0` (and `@12 = 0`, `@16 = 0`) — EXACTLY the `materialize_result_ok`
-    /// convention — so the caller's scope-end flat `DropListStr` frees NOTHING at @12 (it would
-    /// trap on a null `rc_dec` if Ok carried a phantom `len = 1`). The Err arm sets `len@4 = 1`,
-    /// `@12 = msg String`, `@16 tag = 1` (the flat `DropListStr` frees the one owned message). The
-    /// FIFTH host-write sandbox exit, reached only by the self-hosted `fs.write`. Carries
-    /// [`Capability::FsWrite`] — a DISTINCT capability from FsRead (a write is strictly greater
-    /// authority), counted in cap_witness — so a function using it is caps-verified ONLY if it
-    /// declares FsWrite; never accept-but-unsafe. Its dst is a heap Ptr (like [`ReadTextFile`]),
-    /// so the ownership certificate emits an `i` (alloc) for it, balanced by the caller's scope-end
-    /// flat `DropListStr` (sound for BOTH arms given the `len@4 = 0` Ok convention above).
-    WriteTextFile,
-    /// The WASI `path_create_directory` recursive-mkdir sequence, packaged as ONE high-level
-    /// HEAP-RESULT prim — `args = [path]` (a BORROWED `String` handle, the caller still owns
-    /// it), dst = a fresh OWNED `Result[Unit, String]`. Creates the directory at `path`
-    /// (relative to the first preopened dir, leading `/` stripped — the same resolution
-    /// [`WriteTextFile`] uses), creating each missing parent segment (so `a/b/c` makes all
-    /// three); an existing dir (errno EEXIST = 20) counts as success. On success builds
-    /// `Ok(())` (the `len@4 = 0` `materialize_result_ok` convention, IDENTICAL to
-    /// [`WriteTextFile`]'s Ok arm), on a `path_create_directory` error builds
-    /// `Err(<message>)` (`len@4 = 1`, `@12 = msg`, `@16 tag = 1`). A mkdir IS a filesystem
-    /// WRITE, so it REUSES [`Capability::FsWrite`] (NOT a new capability — that would be a
-    /// false distinction); counted in cap_witness exactly like [`WriteTextFile`]. Its dst is
-    /// a heap Ptr, so the ownership certificate emits an `i` (alloc), balanced by the
-    /// caller's scope-end flat `DropListStr` (sound for BOTH arms given the `len@4 = 0` Ok).
-    MakeDir,
-    /// The WASI `path_remove_directory` / `path_unlink_file` RECURSIVE-remove sequence, packaged
-    /// as ONE high-level HEAP-RESULT prim — `args = [path]` (a BORROWED `String` handle, the
-    /// caller still owns it), dst = a fresh OWNED `Result[Unit, String]`. Removes the tree rooted
-    /// at `path` (relative to the first preopened dir, leading `/` stripped — the same resolution
-    /// [`WriteTextFile`] uses): if `path` opens as a directory it RECURSIVELY removes every entry
-    /// (a child directory via `path_remove_directory` after it is emptied, a child file via
-    /// `path_unlink_file`) then removes the now-empty directory; if it is a file it is unlinked
-    /// directly — matching native `fs.remove_all` (`remove_dir_all` for a dir, `remove_file`
-    /// otherwise). On success builds `Ok(())` (the `len@4 = 0` `materialize_result_ok` convention,
-    /// IDENTICAL to [`WriteTextFile`]'s Ok arm), on a removal error builds `Err(<message>)`
-    /// (`len@4 = 1`, `@12 = msg`, `@16 tag = 1`). A remove IS a filesystem WRITE, so it REUSES
-    /// [`Capability::FsWrite`] (NOT a new capability — that would be a false distinction); counted
-    /// in cap_witness exactly like [`WriteTextFile`]. Its dst is a heap Ptr, so the ownership
-    /// certificate emits an `i` (alloc), balanced by the caller's scope-end flat `DropListStr`
-    /// (sound for BOTH arms given the `len@4 = 0` Ok).
-    RemoveAll,
-    /// The WASI `path_filestat_get` existence query, packaged as ONE high-level SCALAR prim —
-    /// `args = [path]` (a BORROWED `String` handle, the caller still owns it), dst = a SCALAR
-    /// `Bool` (an i64 0/1). Stats `path` (relative to the first preopened dir, leading `/`
-    /// stripped — the same resolution [`ReadTextFile`] uses) and yields `1` if a file OR
-    /// directory exists there (errno 0), `0` otherwise — matching native `fs.exists`
-    /// (`std::path::Path::exists`). UNLIKE every other fs prim this is NOT a heap result: a stat
-    /// allocates nothing, so its dst is a plain scalar (NO `materialize_result` block, NO
-    /// scope-end drop, NO ownership-cert `i` — it falls in the scalar-result `_ => {}` arm).
-    /// A stat IS a filesystem READ, so it REUSES [`Capability::FsRead`] (NOT a new capability —
-    /// the SAME accounting as [`ReadTextFile`] → FsRead); counted in cap_witness. Reached only by
-    /// the self-hosted `fs.exists`.
-    PathExists,
-    /// The WASI `path_filestat_get` FULL-stat query — `args = [bufaddr, path]` (a raw scratch
-    /// ADDRESS the caller owns — the self-host's 64-byte Bytes data region — plus a BORROWED
-    /// `String` handle), dst = the SCALAR errno (i64; 0 = the host wrote the 64-byte WASI
-    /// filestat at `bufaddr`: filetype@16, size@32, mtim@48). The self-hosted `fs.stat` reads
-    /// the fields off its own scratch via `prim.load*` and builds the FileStat record in
-    /// ordinary Almide — the prim stays a thin syscall wrapper (no heap result, no ownership
-    /// event; the same scalar-dst discipline as [`PathExists`]). A stat IS a filesystem READ,
-    /// so it REUSES [`Capability::FsRead`] (counted in cap_witness). Reached only by the
-    /// self-hosted `fs.stat`.
-    PathFilestat,
-    /// Release one reference of a RAW heap handle (`(call $rc_dec …)`), the inverse of [`RcInc`].
-    /// The MECHANISM the self-hosted recursive `value.__drop_value` frees a dynamic Value tree with
-    /// (the §4.1-compliant alternative to a hand-written WAT drop): it operates on raw Int handles,
-    /// so its ownership cert is EMPTY (a `Prim` is a no-op in verify_ownership) — like `string_eq`.
-    /// REUSES the proven `$rc_dec` (no new WAT func). args = [addr], no dst (Unit). TRUSTED like the
-    /// inline DropListStr's per-element rc_dec — its leak/double-free safety is the differential
-    /// test's burden (a value.stringify round-trip), NOT the ownership cert. Use is contained to the
-    /// drop routine.
-    RcDec,
-    /// Acquire one reference of a RAW heap handle (`(call $rc_inc …)`) — the self-host `value.array`
-    /// SHALLOW-COPIES a `List[Value]` by `rc_inc`-ing each element into a new owned list (matching
-    /// v0's `items.clone()` observably) so the borrowed `items` param is untouched. args = [addr],
-    /// no dst. REUSES the proven `$rc_inc`. Cert no-op (raw handle), trusted like RcDec.
-    RcInc,
-    /// The FLOAT floor: a `Float` scalar is the i64-uniform value holding the f64 BITS, so
-    /// every float op `reinterpret`s i64→f64, computes, and `reinterpret`s back (a compare /
-    /// to-int yields a real i64). Scalar, no ownership — the cert is untouched (these are
-    /// `Op::Prim`, no-ops in verify_ownership). This opens the whole `float.*` / `math.*`
-    /// f64 category for self-host over `prim.fabs` / `prim.fadd` / `prim.f2i` / etc.
-    FloatUn(FUnOp),
-    FloatBin(FBinOp),
-    /// `float.from_int(x)` — the sitofp floor (#806 step 2): ONE
-    /// `f64.convert_i64_s` (bits-reinterpreted into the i64-uniform float
-    /// slot), replacing the self-host runtime CALL that dominated inlined hot
-    /// loops. `args = [int_value]`, dst = the f64 BITS. A pure scalar
-    /// conversion — no ownership event, no capability.
-    F64FromInt,
-    FloatCmp(FCmpOp),
-    /// `i64.trunc_sat_f64_s(reinterpret(x))` — Float → Int (saturating truncate, v0's `as i64`).
-    FloatToInt,
-    /// `reinterpret(f64.convert_i64_s(x))` — Int → Float.
-    IntToFloat,
-    /// IDENTITY — the raw f64↔i64 BIT reinterpret (`float.to_bits` / `int.bits_to_float`):
-    /// the i64-uniform value ALREADY holds the f64 bits, so this is a no-op pass-through.
-    FloatBits,
-    /// `f32.demote_f64` — Float (f64) → Float32. The narrower f32 value is held as its 32-bit
-    /// pattern in the LOW half of the i64 slot (`i32.reinterpret_f32` then zero-extend). Rounds to
-    /// nearest, matching Rust's `n as f32`.
-    F32Demote,
-    /// `f64.promote_f32` — Float32 → Float (f64). Reads the low-32 f32 pattern (`i32.wrap_i64`
-    /// then `f32.reinterpret_i32`) and widens exactly.
-    F32Promote,
-    /// `f32.convert_i64_s` — Int → Float32 directly (single rounding), matching Rust's `n as f32`.
-    /// Result is the f32 pattern in the low half of the i64 slot.
-    IntToF32,
-    /// IDENTITY — Float32 → its 32-bit pattern as an Int. A Float32 value ALREADY holds the f32
-    /// bits in the low 32 of the i64 slot (high 32 zero, from F32Demote/IntToF32's zero-extend), so
-    /// this is a type-only reinterpret (no-op pass-through), like FloatBits for f64.
-    F32Bits,
-    /// A binary f32 op over two Float32 values (each the low-32 f32 pattern in its i64 slot):
-    /// unwrap → f32 op (per-op f32 rounding, matching native Rust f32 and v0's F32Add family) →
-    /// re-wrap. The f64 `FloatBin` on these bit patterns computed garbage (the low-32 f32 bits
-    /// reinterpreted as f64 are a denormal).
-    F32Bin(FBinOp),
-    /// A binary f32 comparison (`f32.eq`/`lt`/… over the low-32 patterns) → i64-uniform Bool.
-    F32Cmp(FCmpOp),
-    /// A unary f32 op (neg/abs/…) over the low-32 pattern.
-    F32Un(FUnOp),
-}
-
-/// A unary f64 op (the value is the f64 bits in an i64; render reinterprets around it).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum FUnOp {
-    Abs,
-    Sqrt,
-    Floor,
-    Ceil,
-    Neg,
-}
-
-/// A binary f64 op.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum FBinOp {
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Min,
-    Max,
-    /// `f64.copysign(a, b)` — magnitude of `a` with the sign bit of `b` (the basis for an
-    /// exact `f64::signum`: `copysign(1.0, x)`, with NaN handled by the caller).
-    CopySign,
-}
-
-/// An f64 comparison — yields an i64 0/1 (the Bool / `if` condition).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum FCmpOp {
-    Lt,
-    Le,
-    Gt,
-    Ge,
-    Eq,
-    Ne,
-}
-
-/// A scalar integer binary operation.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum IntOp {
-    Add,
-    Sub,
-    Mul,
-    /// Signed division — traps on divide-by-zero (matching v0's checked `DivInt`).
-    Div,
-    /// Signed remainder — traps on divide-by-zero (matching v0's checked `ModInt`).
-    Mod,
-    // Comparisons: produce a Bool scalar (i64 0/1) — the `if` condition. Signed.
-    Lt,
-    Le,
-    Gt,
-    Ge,
-    Eq,
-    Ne,
-    // Bitwise i64 ops (the int.band/bor/bxor/bshl/bshr floor). Scalar, no ownership.
-    And,
-    Or,
-    Xor,
-    Shl,
-    /// Arithmetic (sign-extending) shift right, matching v0's `>>` on `i64`.
-    Shr,
-    /// LOGICAL (zero-filling) shift right (`i64.shr_u`) — for unsigned/bit-width ops like
-    /// int.rotate_* which shift the value as a u64. The shift amount is wasm-masked to 0..63.
-    ShrU,
-}
-
-/// A runtime function the MIR can call. An enum (not a string) so the renderer
-/// mapping is TOTAL and the runtime surface is a closed, auditable set.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum RtFn {
-    /// `list[index] = value` in place (after a [`Op::MakeUnique`]).
-    ListSet,
-    /// push a value onto a list in place (after a [`Op::MakeUnique`]); the
-    /// result is rebound to `dst` (the buffer may move).
-    ListPush,
-    /// `println` a list as `label=e0,e1,…`.
-    PrintList,
-    /// `println` a scalar integer.
-    PrintInt,
-    /// `println` a heap string (the value-semantics subset's string print). A
-    /// WITNESS-LEVEL primitive today: it carries the ownership (borrows the
-    /// string handle) and capability ([`Capability::Stdout`]) facts the proven
-    /// checker re-verifies, but the renderers do NOT lower it yet — strings are
-    /// `Init::Opaque` skeletons in this subset (no content bytes), so a faithful
-    /// `print_str` render awaits the string-content lowering brick. Until then a
-    /// renderer asked to emit it refuses LOUDLY (the catch-all panic), never
-    /// silently — the flight-grade totality rule.
-    PrintStr,
-}
-
-impl RtFn {
-    /// The host [`Capability`] this runtime function reaches, if any. Pure heap
-    /// ops touch no host effect; the print ops reach [`Capability::Stdout`]. This
-    /// is the SINGLE mapping the capability witness derives "used capabilities"
-    /// from — exhaustive, so a new effectful runtime fn cannot silently escape
-    /// the sandbox accounting.
-    pub const fn capability(self) -> Option<Capability> {
-        match self {
-            RtFn::ListSet | RtFn::ListPush => None,
-            RtFn::PrintList | RtFn::PrintInt | RtFn::PrintStr => Some(Capability::Stdout),
-        }
-    }
-}
-
-/// A host CAPABILITY a function may reach — the unit of the sandbox promise
-/// (the 4th flight-grade property, proofs/CapabilityBound.v: a program reaches
-/// ONLY the capabilities it declares). A VALUE OBJECT, not a raw id: you write
-/// `Capability::Stdout`, never `0`. The stable registry id the proven checker
-/// compares is recovered via [`Capability::id`], so the "Stdout = 0" mapping
-/// lives in exactly ONE place and MUST match the Coq capability registry. The
-/// set is closed and grows only as the runtime gains host effects (fs, net, …).
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash, PartialOrd, Ord)]
-pub enum Capability {
-    /// Writing to standard output (the only host effect the current MIR subset
-    /// reaches, via [`RtFn::PrintInt`] / [`RtFn::PrintList`]).
-    Stdout,
-    /// Reading host ENTROPY — the WASI `random_get` floor ([`PrimKind::RandomGet`]),
-    /// reached by the self-hosted `random.int`. The second sandbox exit. A pure `fn`
-    /// declares ∅, so it can NEVER reach entropy un-witnessed (the checker REJECTS
-    /// `used ⊄ allowed`); only an `effect fn` (which declares the host caps) may.
-    Entropy,
-    /// Reading the program's COMMAND-LINE ARGUMENTS — the WASI `args_sizes_get` /
-    /// `args_get` floor ([`PrimKind::ArgsGetList`]), reached by the self-hosted
-    /// `env.args`. The third sandbox exit. Accounted exactly like Entropy/Stdout: a
-    /// pure `fn` declares ∅ and so can NEVER read argv un-witnessed (the checker
-    /// REJECTS `used ⊄ allowed`); only an `effect fn` (which declares the host caps) may.
-    CliArgs,
-    /// Reading a FILE from the host filesystem — the WASI `path_open` / `fd_read` floor
-    /// ([`PrimKind::ReadTextFile`]), reached by the self-hosted `fs.read_text`. The fourth
-    /// sandbox exit. Accounted exactly like CliArgs/Entropy/Stdout: a pure `fn` declares ∅
-    /// and so can NEVER read a file un-witnessed (the checker REJECTS `used ⊄ allowed`);
-    /// only an `effect fn` (which declares the host caps) may.
-    FsRead,
-    /// Writing a FILE to the host filesystem — the WASI `path_open(O_CREAT|O_TRUNC)` /
-    /// `fd_write` floor ([`PrimKind::WriteTextFile`]), reached by the self-hosted `fs.write`.
-    /// The fifth sandbox exit. A STRICTLY GREATER authority than [`Self::FsRead`] (a write
-    /// creates/truncates host state), so it is a DISTINCT capability with its own id — never
-    /// aliased to FsRead (conflating read and write would be a capability lie: a fn declaring
-    /// only read could mutate the filesystem). Accounted exactly like FsRead: a pure `fn`
-    /// declares ∅ and so can NEVER write a file un-witnessed (the checker REJECTS
-    /// `used ⊄ allowed`); only an `effect fn` (which declares the host caps) may.
-    FsWrite,
-    /// Reading the host WALL CLOCK — the WASI `clock_time_get` floor
-    /// ([`PrimKind::ClockTimeGet`]), reached by the self-hosted `env.unix_timestamp`. The
-    /// sixth sandbox exit. A clock read is neither a filesystem effect nor an entropy draw,
-    /// so it is a DISTINCT capability with its own id — never aliased to FsRead/FsWrite or
-    /// Entropy. Accounted exactly like Entropy/FsRead: a pure `fn` declares ∅ and so can
-    /// NEVER read the clock un-witnessed (the checker REJECTS `used ⊄ allowed`); only an
-    /// `effect fn` (which declares the host caps) may.
-    Clock,
-    /// Reading STANDARD INPUT — the WASI `fd_read`-from-fd-0 floor ([`PrimKind::ReadLine`]),
-    /// reached by the self-hosted `io.read_line`. The seventh sandbox exit. Reading stdin is
-    /// neither a write, a filesystem read, an entropy draw, nor a clock read, so it is a DISTINCT
-    /// capability with its own id — never aliased to FsRead/FsWrite/Entropy/Clock (a fn that
-    /// consumes the operator's input stream is a real, separately-grantable authority). Accounted
-    /// exactly like Entropy/FsRead: a pure `fn` declares ∅ and so can NEVER read stdin
-    /// un-witnessed (the checker REJECTS `used ⊄ allowed`); only an `effect fn` (which declares
-    /// the host caps) may.
-    Stdin,
-}
-
-impl Capability {
-    /// The stable registry id — the ONLY place a `Capability` becomes a number.
-    /// proofs/CapabilityBound.v's checker is GENERIC over `list nat` (a `subset_check`,
-    /// no per-capability enumeration), so it needs no edit to admit a new id — only
-    /// this mapping must stay injective + stable (Stdout = 0, Entropy = 1, CliArgs = 2,
-    /// FsRead = 3, FsWrite = 4, Clock = 5, Stdin = 6).
-    pub const fn id(self) -> u32 {
-        match self {
-            Capability::Stdout => 0,
-            Capability::Entropy => 1,
-            Capability::CliArgs => 2,
-            Capability::FsRead => 3,
-            Capability::FsWrite => 4,
-            Capability::Clock => 5,
-            Capability::Stdin => 6,
-        }
-    }
-}
-
-/// A wasm IMPORT-signature value type — the host-facing valtype an
-/// [`Op::CallImport`] argument/result is mapped to from its declared Almide type
-/// (Int→`I64`, Float→`F64`, Bool→`I32`, String/heap pointer→`I32`). The MIR is
-/// i64-uniform for scalars (a Float local holds the f64 BITS) and i32 for heap
-/// handles, so the render coerces each local to/from this valtype at the call.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum WasmAbi {
-    /// A 64-bit integer — the MIR scalar local passes through directly.
-    I64,
-    /// A 64-bit float — the MIR i64 local holds its bits; reinterpret around the call.
-    F64,
-    /// A 32-bit integer — a heap pointer (MIR i32, direct) or a Bool (MIR i64, wrapped).
-    I32,
-}
-
-impl WasmAbi {
-    /// The WAT valtype keyword for an import signature.
-    pub fn wat(self) -> &'static str {
-        match self {
-            WasmAbi::I64 => "i64",
-            WasmAbi::F64 => "f64",
-            WasmAbi::I32 => "i32",
-        }
-    }
-}
-
-/// An argument to a runtime [`Op::Call`] / user [`Op::CallFn`].
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum CallArg {
-    /// A heap handle (borrowed by the call — live-checked, refcount unchanged).
-    Handle(ValueId),
-    /// A scalar value (a `ValueId` of scalar Repr — no ownership).
-    Scalar(ValueId),
-    /// An immediate integer (index / value).
-    Imm(i64),
-    /// An immediate string (a print label).
-    Label(String),
-}
-
-/// A function parameter: a value the caller supplies, with its [`Repr`]. A heap
-/// param is BORROWED (the v1 calling convention): the CALLER retains ownership
-/// and releases it; the callee gets a live handle but no owned reference. So a
-/// param contributes NO `+1` to the ownership certificate — an owned-param `+1`
-/// would be synthetic (no runtime `Alloc`/`rc_inc` backs it), the gate-blind
-/// use-after-free class. A body that needs to consume or return a param must
-/// first `Dup` it (acquire its own reference). A scalar param carries no
-/// ownership. (Per-param move-mode signatures are a later refinement.)
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct MirParam {
-    pub value: ValueId,
-    pub repr: Repr,
-}
-
-/// A MIR function: params, a flat ownership-explicit op sequence, and an
-/// optional returned value (moved out — a [`Op::Consume`] of `ret` is implied at
-/// the boundary).
-#[derive(Clone, Debug, PartialEq, Eq, Default)]
-pub struct MirFunction {
-    pub name: String,
-    pub params: Vec<MirParam>,
-    pub ops: Vec<Op>,
-    pub ret: Option<ValueId>,
-    /// The host [`Capability`]s this function is PERMITTED to reach (its effect
-    /// signature, lowered). The capability witness checks the capabilities the
-    /// body actually uses against this declared bound — accept ⟹ no undeclared
-    /// host effect (proofs/CapabilityBound.v). Empty = a pure/sandboxed function.
-    pub declared_caps: Vec<Capability>,
-    /// RENDER-ONLY side table: a value → the i64-SLOT INDICES that hold an OWNED heap
-    /// handle, for a MIXED scalar+heap record/tuple block (e.g. `R { name: String, n: Int }`
-    /// = `[0]`). It refines the recursive free of an [`Op::DropListStr`] on such a value:
-    /// instead of the uniform "free EVERY slot" loop (correct only for a homogeneous
-    /// `List[String]`), the render frees exactly these slots, then the block. A value
-    /// ABSENT from this table keeps the uniform-loop behavior (`List[String]` / all-heap
-    /// aggregate). This carries NO ownership semantics — the certificate sees a `DropListStr`
-    /// as the SAME single `d` regardless (each heap field was already accounted `m`/consumed
-    /// at its move-in store), exactly as for `List[String]`. So it is a pure rendering
-    /// refinement (like the `DropValue` tag dispatch) — NOT a new op or certificate event.
-    pub heap_slot_masks: BTreeMap<ValueId, Vec<usize>>,
-}
-
-/// A whole MIR program.
-#[derive(Clone, Debug, PartialEq, Eq, Default)]
-pub struct MirProgram {
-    pub functions: Vec<MirFunction>,
-    /// `pub fn` export roots (#457 — the fns the v0 emitter also exports). Each entry:
-    /// (export_name — the `@export(wasm, "sym")` override or the fn name, internal fn
-    /// name, per-param is_float, ret: None = void / Some(is_float)). A Float-bearing
-    /// signature renders through a thin `f64.reinterpret_i64` wrapper so the export
-    /// presents REAL f64s (the v0 ABI) while the internal fn keeps the i64-bits
-    /// convention. Populated by the pipeline from the MAIN program's Public non-test
-    /// non-generic functions; empty everywhere else.
-    pub exports: Vec<(String, String, Vec<bool>, Option<bool>)>,
-    /// The number of MUTABLE module-level `var` storage slots. Slot `i` lives at linear
-    /// address [`mg_slot_addr`]`(i)` — the 8-byte region `[MG_SLOT_BASE, MG_SLOT_BASE +
-    /// 8*count)` carved between the print line buffer (which ends at `MG_SLOT_BASE`) and
-    /// the bump allocator (whose base the renderer shifts to `MG_SLOT_BASE + 8*count`).
-    /// A count of 0 renders byte-identically to a program with no mutable globals.
-    pub mutable_global_count: u32,
-}
-
-/// The base linear-memory address of the mutable-global slot region (== the renderer's
-/// `HEAP_BASE`; with no mutable globals the bump allocator starts exactly here).
-pub const MG_SLOT_BASE: u32 = 8192;
-
-/// The linear-memory address of mutable-global slot `index` (one uniform 8-byte slot per
-/// module-level `var`: a scalar holds its i64 value, a heap global its block handle).
-pub const fn mg_slot_addr(index: u32) -> u32 {
-    MG_SLOT_BASE + 8 * index
-}
-
-// ─────────────────────────── Ownership verifier ───────────────────────────
-//
-// The executable ownership invariant (#575/#576). A symbolic refcount
-// interpretation over the ops: every heap value's owner count must return to 0
-// (every reference dropped or moved out), never go negative (double-free), and
-// never be used after it reaches 0 / is moved (use-after-free / -move).
-
-/// What an ownership violation is.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum ViolationKind {
-    /// A `drop` of a value whose owner count is already 0.
-    DoubleFree,
-    /// A `dup`/`borrow`/`make_unique`/`pure`-use of a freed value.
-    UseAfterFree,
-    /// A `consume` of a value already moved out (count 0).
-    UseAfterMove,
-    /// A heap value still owned (count > 0) at function end.
-    Leak,
-    /// The two arms of an `IfThen`/`Else`/`EndIf` branch leave an object at
-    /// DIFFERENT owner counts — whichever way the branch goes at runtime, the
-    /// later accounting is wrong for the other path (a path-dependent leak or
-    /// double-free). Mirrors the proven checker's `CBranch` agreement rule.
-    BranchDisagreement,
-}
-
-/// A located ownership violation.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub struct Violation {
-    /// Index into `func.ops`; equals `ops.len()` for an end-of-function leak.
-    pub op_index: usize,
-    pub value: ValueId,
-    pub kind: ViolationKind,
-}
-
-/// Verify the ownership invariant for one function.
-///
-/// Returns `Ok(())` if the MIR is balanced (the by-construction guarantee the
-/// renderers rely on), or every violation found (deterministic order). This is
-/// the MIR-level analogue of the Perceus belt's IR check, but it is the SINGLE
-/// source — there is no second hand-written copy in a renderer to drift from.
-/// The mutable scan state of [`verify_ownership`] — one step per op (#781:
-/// the cog-140 loop body became [`OwnershipScan::step`]).
-struct OwnershipScan {
-    object_of: BTreeMap<ValueId, ValueId>,
-    rc: BTreeMap<ValueId, i64>,
-    dead: BTreeMap<ValueId, bool>,
-    borrowed: BTreeSet<ValueId>,
-    branches: Vec<BranchFrame>,
-    violations: Vec<Violation>,
-}
-
-    struct BranchFrame {
-        entry_rc: BTreeMap<ValueId, i64>,
-        entry_dead: BTreeMap<ValueId, bool>,
-        then_exit: Option<(BTreeMap<ValueId, i64>, BTreeMap<ValueId, bool>)>,
-    }
-
-impl OwnershipScan {
-    /// One op's ownership transition. Verbatim text move of the scan loop body
-    /// (locals renamed to fields).
-    fn step(&mut self, i: usize, op: &Op) {
-        match op {
-            Op::Alloc { dst, repr, .. } => {
-                debug_assert!(repr.is_heap(), "Alloc of a non-heap repr is malformed MIR");
-                self.object_of.insert(*dst, *dst);
-                self.rc.insert(*dst, 1);
-                self.dead.insert(*dst, false);
-            }
-            // A rung-4 scalar-list LITERAL is alloc-class: one fresh owned object
-            // (the identical accounting the replaced `Alloc{DynList}` had). Its
-            // element values are raw i64 slot scalars — no ownership to check.
-            Op::ListLit { dst, .. } => {
-                self.object_of.insert(*dst, *dst);
-                self.rc.insert(*dst, 1);
-                self.dead.insert(*dst, false);
-            }
-            // The rung-4 element load/store BORROW the list handle (live-check,
-            // no refcount change — exactly the `Borrow`/`MakeUnique` discipline);
-            // the scalar element/index/value carry no ownership.
-            Op::ListGetScalar { list, .. } | Op::ListSetScalar { list, .. } => {
-                if live_object(&self.object_of, &self.rc, &self.dead, &self.borrowed, *list).is_none() {
-                    self.violations.push(violation(i, *list, ViolationKind::UseAfterFree));
-                }
-            }
-            Op::Const { dst: _ } | Op::ConstInt { .. } => {
-                // A scalar — no ownership accounting.
-            }
-            Op::FuncRef { .. } => {
-                // A function-table slot index — a scalar constant, no ownership.
-            }
-            Op::Dup { dst, src } => {
-                if let Some(o) = live_object(&self.object_of, &self.rc, &self.dead, &self.borrowed, *src) {
-                    // Acquire OUR own reference. A `Dup` of a self.borrowed param has no
-                    // prior self.rc entry (we owned none) — start it at 0, then +1.
-                    *self.rc.entry(o).or_insert(0) += 1;
-                    self.object_of.insert(*dst, o);
-                    self.dead.insert(*dst, false);
-                } else {
-                    self.violations.push(violation(i, *src, ViolationKind::UseAfterFree));
-                }
-            }
-            // A `DropListStr`/`DropListValue` releases the LIST object exactly like a `Drop` (the
-            // recursive element free is a RENDER concern, gated on self.rc==1; the cert sees one −1 on the
-            // list — its elements were `Consume`d into it when stored).
-            Op::Drop { v }
-            | Op::DropListStr { v }
-            | Op::DropValue { v }
-            | Op::DropListValue { v }
-            | Op::DropListStrValue { v }
-            | Op::DropListStrStr { v }
-            | Op::DropListIntStr { v }
-            | Op::DropListStrInt { v }
-            | Op::DropResultListValue { v }
-            | Op::DropResultValue { v }
-            | Op::DropResultStrInt { v }
-            | Op::DropResultValueInt { v }
-            | Op::DropResultListValueInt { v }
-            | Op::DropResultListStrInt { v }
-            | Op::DropResultListStr { v }
-            | Op::DropListListStr { v }
-            | Op::DropVariant { v, .. }
-            | Op::DropWrapperRec { v, .. } => {
-                match release(&self.object_of, &mut self.rc, &mut self.dead, &self.borrowed, *v) {
-                    Ok(()) => {}
-                    Err(()) => self.violations.push(violation(i, *v, ViolationKind::DoubleFree)),
-                }
-            }
-            Op::Consume { v } => match release(&self.object_of, &mut self.rc, &mut self.dead, &self.borrowed, *v) {
-                Ok(()) => {}
-                Err(()) => self.violations.push(violation(i, *v, ViolationKind::UseAfterMove)),
-            },
-            Op::Borrow { v } | Op::MakeUnique { v } => {
-                if live_object(&self.object_of, &self.rc, &self.dead, &self.borrowed, *v).is_none() {
-                    self.violations.push(violation(i, *v, ViolationKind::UseAfterFree));
-                }
-            }
-            Op::Pure { dst: _, uses } => {
-                for v in uses {
-                    // Only heap handles are accountable; scalar uses are absent
-                    // from `self.object_of` and correctly skipped.
-                    if self.object_of.contains_key(v)
-                        && live_object(&self.object_of, &self.rc, &self.dead, &self.borrowed, *v).is_none()
-                    {
-                        self.violations.push(violation(i, *v, ViolationKind::UseAfterFree));
-                    }
-                }
-            }
-            // A runtime/user call BORROWS its heap-handle args (live-checked, no
-            // refcount change). Immediate/label args carry no ownership. A call
-            // whose `result` is a heap repr returns a FRESH OWNED value (the
-            // callee allocated and moved it out — the return-mode signature): the
-            // `dst` becomes a new owned object, like Alloc.
-            Op::Call { args, dst, result, .. }
-            | Op::CallFn { args, dst, result, .. }
-            // A CallImport (a host wasm import) has the SAME ownership shape: heap-handle
-            // args are BORROWED, a heap result is a FRESH OWNED value (the host returns a
-            // pointer the caller now owns). Its scalar args carry no ownership.
-            | Op::CallImport { args, dst, result, .. }
-            // A CallIndirect has the same ownership shape as a CallFn: its heap-arg handles
-            // must be live, a heap result is a FRESH OWNED value. The `table_idx` is a
-            // scalar closure value (no ownership).
-            | Op::CallIndirect { args, dst, result, .. } => {
-                for a in args {
-                    if let CallArg::Handle(v) = a {
-                        if live_object(&self.object_of, &self.rc, &self.dead, &self.borrowed, *v).is_none() {
-                            self.violations.push(violation(i, *v, ViolationKind::UseAfterFree));
-                        }
-                    }
-                }
-                if let (Some(d), Some(r)) = (dst, result) {
-                    if r.is_heap() {
-                        self.object_of.insert(*d, *d);
-                        self.rc.insert(*d, 1);
-                        self.dead.insert(*d, false);
-                    }
-                }
-            }
-            // The if-markers carry no ownership of their own, but they scope the
-            // BRANCH JOIN: both arms run from the entry state and must agree.
-            Op::IfThen { .. } => {
-                self.branches.push(BranchFrame {
-                    entry_rc: self.rc.clone(),
-                    entry_dead: self.dead.clone(),
-                    then_exit: None,
-                });
-            }
-            Op::Else { .. } => {
-                if let Some(fr) = self.branches.last_mut() {
-                    fr.then_exit = Some((self.rc.clone(), self.dead.clone()));
-                    self.rc = fr.entry_rc.clone();
-                    self.dead = fr.entry_dead.clone();
-                }
-            }
-            Op::EndIf { .. } => {
-                if let Some(fr) = self.branches.pop() {
-                    let (then_rc, then_dead) = match fr.then_exit {
-                        Some(t) => t,
-                        // No Else marker: everything since IfThen was the then arm;
-                        // the else arm is empty (= the entry state).
-                        None => {
-                            let cur = (self.rc.clone(), self.dead.clone());
-                            self.rc = fr.entry_rc.clone();
-                            self.dead = fr.entry_dead.clone();
-                            cur
-                        }
-                    };
-                    // Agreement per object (absent = 0 owned refs).
-                    let keys: BTreeSet<ValueId> =
-                        then_rc.keys().chain(self.rc.keys()).copied().collect();
-                    for k in keys {
-                        let a = then_rc.get(&k).copied().unwrap_or(0);
-                        let b = self.rc.get(&k).copied().unwrap_or(0);
-                        if a != b {
-                            self.violations.push(violation(i, k, ViolationKind::BranchDisagreement));
-                        }
-                    }
-                    // Continue with the JOIN: pointwise max keeps the run stable
-                    // after a reported disagreement (no cascading underflows); on
-                    // agreement it is the common value. A handle self.dead on EITHER
-                    // path is unusable after the merge.
-                    for (k, v) in then_rc {
-                        let e = self.rc.entry(k).or_insert(0);
-                        if v > *e {
-                            *e = v;
-                        }
-                    }
-                    for (k, d) in then_dead {
-                        let e = self.dead.entry(k).or_insert(d);
-                        *e = *e || d;
-                    }
-                }
-            }
-            // Scalar arithmetic — no ownership.
-            // A scalar arithmetic op and a primitive-floor op carry no ownership: a
-            // scalar result is Copy and a `Prim` handle arg is BORROWED (read only).
-            Op::IntBinOp { .. }
-            // Loop markers carry no ownership; the body ops between them are
-            // per-iteration-balanced (verified flat, one iteration).
-            | Op::LoopStart
-            | Op::LoopBreakUnless { .. }
-            | Op::LoopEnd => {}
-            // VALUE-RC modeling (柱C extension) — bring the Value refcount ops out of the prim blind
-            // spot for the NAMEABLE case: prim.handle(v) carries its source object in args[0], so the
-            // self.rc events on it verify against the same self.rc machine. load64-fed handles have no carrier
-            // and stay unmodeled (the differential-test floor). MIRRORED in ownership_certificate.
-            Op::Prim { kind, dst, args } => match kind {
-                PrimKind::Handle => {
-                    if let (Some(d), Some(&o)) =
-                        (dst.as_ref(), args.first().and_then(|a| self.object_of.get(a)))
-                    {
-                        self.object_of.insert(*d, o);
-                    }
-                }
-                PrimKind::RcInc => {
-                    if let Some(&o) = args.first().and_then(|a| self.object_of.get(a)) {
-                        *self.rc.entry(o).or_insert(0) += 1;
-                    }
-                }
-                PrimKind::RcDec => {
-                    if let Some(&o) = args.first().and_then(|a| self.object_of.get(a)) {
-                        if self.rc.get(&o).copied().unwrap_or(0) >= 1 {
-                            *self.rc.entry(o).or_insert(0) -= 1;
-                        }
-                    }
-                }
-                _ => {}
-            },
-            // `SetLocal` into a HEAP slot is a loop-carried REBIND (`acc = acc + [x]`):
-            // the slot now aliases the source's object. The slot's OLD object was
-            // released by a preceding `Drop` in the loop body, so rebinding makes the
-            // slot LIVE again (= the new object), preserving the per-iteration invariant
-            // (slot owns exactly one ref at the body's start and end) — exactly the
-            // soundness condition proved in OwnershipChecker.v's `check_line_unroll_sound`
-            // (a self.rc-preserving loop body is leak/double-free-free for any iteration
-            // count). For a SCALAR src (the scalar-TCO loop var) `self.object_of` has no
-            // entry, so this is a no-op, as before.
-            Op::SetLocal { local, src } => {
-                if let Some(o) = self.object_of.get(src).copied() {
-                    self.object_of.insert(*local, o);
-                    self.dead.insert(*local, false);
-                }
-            }
-        }
-    }
-}
-
-pub fn verify_ownership(func: &MirFunction) -> Result<(), Vec<Violation>> {
-    // Handle ≠ object. Each known heap HANDLE (ValueId) maps to its OBJECT (the
-    // `Alloc`'d representative ValueId); the refcount is per OBJECT. A handle is
-    // also tracked LIVE/dead, so a use of a handle after its own drop/consume is
-    // caught even when the object lives on through a sibling handle.
-    let mut object_of: BTreeMap<ValueId, ValueId> = BTreeMap::new();
-    let mut rc: BTreeMap<ValueId, i64> = BTreeMap::new(); // keyed by object — OUR (callee's) owned refs
-    let mut dead: BTreeMap<ValueId, bool> = BTreeMap::new(); // keyed by handle
-    let mut violations: Vec<Violation> = Vec::new();
-
-    // Heap params are BORROWED by default (the v1 calling convention): the CALLER
-    // owns the reference and releases it at its own scope end; the callee gets a
-    // LIVE handle but holds NO owned reference of its own (its rc starts at 0).
-    // This is the exact dual of the certificate omitting the param's `i` event —
-    // an owned-param `+1` would be SYNTHETIC (no `Alloc`/`rc_inc` backs it), the
-    // gate-blind use-after-free class. A body that wants to consume or return a
-    // param must first `Dup` it (acquire its own ref); a release with rc 0 (the
-    // `borrowed` object, never `Dup`'d) fails — exactly the cert's `d`/`m` at
-    // rc 0, which the proven checker faults.
-    let mut borrowed: BTreeSet<ValueId> = BTreeSet::new();
-    for p in &func.params {
-        if p.repr.is_heap() {
-            object_of.insert(p.value, p.value);
-            dead.insert(p.value, false);
-            borrowed.insert(p.value);
-        }
-    }
-
-    // BRANCH JOIN (mirrors the proven checker's `CBranch` rule): each arm of an
-    // `IfThen`/`Else`/`EndIf` runs from the SAME entry state, and the arms must
-    // AGREE on every object's leaving count (the net may be nonzero — a
-    // heap-result branch nets +1 through either arm). Folding the arms FLAT
-    // (the old model) counted BOTH arms' events, silently accepting cross-arm
-    // compensation — a `Consume` in one arm "balancing" the other arm's missing
-    // release, i.e. a path-dependent leak/double-free.
-
-    // Decomposed (#781, cog 140): the per-op transition lives in
-    // `OwnershipScan::step`; the maps moved into the scan struct verbatim.
-    let mut scan = OwnershipScan {
-        object_of,
-        rc,
-        dead,
-        borrowed,
-        branches: Vec::new(),
-        violations,
-    };
-    for (i, op) in func.ops.iter().enumerate() {
-        scan.step(i, op);
-    }
-    let OwnershipScan { object_of, mut rc, mut dead, borrowed, mut violations, .. } = scan;
-
-    // A heap return value is MOVED OUT to the caller. It must be a reference WE
-    // own (an `Alloc`/call-result, or a `Dup` we acquired): releasing it transfers
-    // our reference out. Returning a BORROWED param we never acquired (rc 0) would
-    // give the caller a SECOND owner of the caller's own reference — a double-free.
-    // `release` fails there (rc 0) and we record it, the dual of the cert's `m` at
-    // rc 0 which the proven checker faults.
-    if let Some(r) = func.ret {
-        if object_of.contains_key(&r)
-            && release(&object_of, &mut rc, &mut dead, &borrowed, r).is_err()
-        {
-            violations.push(violation(func.ops.len(), r, ViolationKind::UseAfterMove));
-        }
-    }
-
-    // Leak check: every object's references must have left (dropped or moved).
-    for (o, c) in &rc {
-        if *c > 0 {
-            violations.push(violation(func.ops.len(), *o, ViolationKind::Leak));
-        }
-    }
-
-    if violations.is_empty() {
-        Ok(())
-    } else {
-        Err(violations)
-    }
-}
-
-fn violation(op_index: usize, value: ValueId, kind: ViolationKind) -> Violation {
-    Violation { op_index, value, kind }
-}
-
-/// The object a handle denotes, iff the handle is live. A handle is live when it
-/// is not yet dropped AND either WE hold a reference to its object (rc ≥ 1) OR
-/// the object is a `borrowed` param the CALLER keeps alive for the call's
-/// duration (a borrow is always valid against the caller's reference, even when
-/// our own count is 0). `None` = dead/unknown handle, or a non-borrowed object
-/// whose references have all left.
-fn live_object(
-    object_of: &BTreeMap<ValueId, ValueId>,
-    rc: &BTreeMap<ValueId, i64>,
-    dead: &BTreeMap<ValueId, bool>,
-    borrowed: &BTreeSet<ValueId>,
-    v: ValueId,
-) -> Option<ValueId> {
-    if dead.get(&v).copied().unwrap_or(true) {
-        return None; // unknown handle or already dropped/consumed
-    }
-    let o = *object_of.get(&v)?;
-    if borrowed.contains(&o) || rc.get(&o).copied().unwrap_or(0) >= 1 {
-        Some(o)
-    } else {
-        None
-    }
-}
-
-/// Release one reference held by handle `v` (drop or consume): mark the handle
-/// dead and decrement OUR object's refcount. `Err(())` if `v` is not live, OR if
-/// we hold no reference of our own to release (rc 0 — e.g. a `borrowed` param we
-/// never `Dup`'d): freeing a reference we do not own is a double-free against the
-/// caller, so it is rejected rather than silently underflowed.
-fn release(
-    object_of: &BTreeMap<ValueId, ValueId>,
-    rc: &mut BTreeMap<ValueId, i64>,
-    dead: &mut BTreeMap<ValueId, bool>,
-    borrowed: &BTreeSet<ValueId>,
-    v: ValueId,
-) -> Result<(), ()> {
-    match live_object(object_of, rc, dead, borrowed, v) {
-        Some(o) if rc.get(&o).copied().unwrap_or(0) >= 1 => {
-            *rc.get_mut(&o).expect("a held reference has a refcount") -= 1;
-            dead.insert(v, true);
-            Ok(())
-        }
-        _ => Err(()),
-    }
-}
-
-include!("lib_p2.rs");
+include!("lib_b.rs");
+include!("lib_c.rs");
